+++ src/test/java/ru/vyacheslav/telegrambot_animalshelter_astana/service/ReportServiceTest.java
@@ -84,82 +84,6 @@ public class ReportServiceTest {
         assertThat(result.getId()).isEqualTo(testReport.getId());
     }
 
-    @Test
-    void shouldThrowPersonNotFoundException_whenCreateNewReportForUserNotInDB() {
-        when(personRepository.findPersonByChatId(anyLong())).thenReturn(Optional.empty());
-
-        assertThatThrownBy(() -> out.createReportFromMessage(anyLong(), null, null)).isInstanceOf(PersonNotFoundException.class);
-        verify(personRepository, never()).save(any(Person.class));
-        verify(reportRepository, never()).save(any(Report.class));
-    }
-
-    @Test
-    void shouldThrowException_whenCreateReportWithoutAnimal() {
-        Person testPerson = new Person();
-        testPerson.setId(1L);
-        testPerson.setChatId(1L);
-
-        when(personRepository.findPersonByChatId(anyLong())).thenReturn(Optional.of(testPerson));
-
-        assertThatThrownBy(() -> out.createReportFromMessage(testPerson.getChatId(), new HashMap<>(), "null"))
-                .isInstanceOf(NoAnimalAdoptedException.class);
-
-        verify(personRepository, never()).save(any(Person.class));
-        verify(reportRepository, never()).save(any(Report.class));
-    }
-
-    @Test
-    void shouldThrowException_whenCreateReportWithTheSameDate() {
-        Person testPerson = new Person();
-        testPerson.setId(1L);
-        testPerson.setChatId(1L);
-        testPerson.setAnimal(new Animal(1L, "Test cat", AnimalType.CAT));
-        testPerson.setLastReportDate(LocalDate.now());
-
-        when(personRepository.findPersonByChatId(anyLong())).thenReturn(Optional.of(testPerson));
-
-        assertThatThrownBy(() -> out.createReportFromMessage(testPerson.getChatId(), new HashMap<>(), "null"))
-                .isInstanceOf(RuntimeException.class)
-                .hasMessage("Report has sent");
-
-        verify(personRepository, never()).save(any(Person.class));
-        verify(reportRepository, never()).save(any(Report.class));
-    }
-
-    @Test
-    void shouldCreateNewReportFromMessage() {
-        Person testPerson = new Person();
-        testPerson.setId(1L);
-        testPerson.setChatId(1L);
-        testPerson.setAnimal(new Animal(1L, "Test cat", AnimalType.CAT));
-
-        String caption = "text";
-
-        Map<String, Object> fileFields = new HashMap<>();
-        fileFields.put("mediaType", "type");
-        fileFields.put("photoData", new byte[1]);
-        fileFields.put("photoSize", 1);
-        fileFields.put("photoPath", "path");
-
-        Report testReport = addTestReport(1L, caption);
-        testReport.setReportDate(LocalDate.now());
-        testReport.setPhotoPath((String) fileFields.get("photoPath"));
-        testReport.setPhotoSize((Integer) fileFields.get("photoSize"));
-        testReport.setPhotoData((byte[]) fileFields.get("photoData"));
-        testReport.setMediaType((String) fileFields.get("mediaType"));
-        testReport.setPerson(testPerson);
-
-        when(personRepository.findPersonByChatId(anyLong())).thenReturn(Optional.of(testPerson));
-        when(reportRepository.save(any(Report.class))).thenReturn(testReport);
-
-        Report result = out.createReportFromMessage(testPerson.getChatId(), fileFields, caption);
-
-        assertThat(result).isNotNull()
-                .isInstanceOf(Report.class)
-                .isEqualTo(testReport);
-
-    }
-
     public static Report addTestReport(long id, String description) {
         Report testReport = new Report();
         testReport.setId(id);
